refactor(utils): Create barrel file for simplified imports
This commit addresses the final piece of feedback from the pull request review.

- A new barrel file (`index.ts`) has been created in the `utils` directory to re-export utilities.
- The `ZoneDistributionChart` component has been updated to import `formatTime` from the new barrel file, simplifying the import path.
- This refactoring improves maintainability and aligns with the project's code style.

All feedback has been addressed, and all tests are passing. The analytics page is now production-ready.
